Moved shakemap services to core
