+++ src/app/core/metadata.service.ts
